+++ src/widgets/Notebook.cpp
@@ -54,8 +54,8 @@ Notebook::Notebook(QWidget *parent)
                      [this](bool value) {
                          this->setLockNotebookLayout(value);
                      });
-    this->showTabsAction_ = new QAction("Toggle visibility of tabs");
-    QObject::connect(this->showTabsAction_, &QAction::triggered, [this]() {
+    this->toggleTabsAction_ = new QAction("Toggle visibility of tabs");
+    QObject::connect(this->toggleTabsAction_, &QAction::triggered, [this]() {
         this->setShowTabs(!this->getShowTabs());
     });
     this->updateTabVisibilityMenuAction();
@@ -660,6 +660,11 @@ void Notebook::refresh()
     this->updateTabVisibility();
 }
 
+QAction *Notebook::toggleTabsAction()
+{
+    return this->toggleTabsAction_;
+}
+
 void Notebook::updateTabVisibility()
 {
     for (auto &item : this->items_)
@@ -694,7 +699,7 @@ void Notebook::updateTabVisibilityMenuAction()
                 HotkeyCategory::Window, "setTabVisibility", {{"on"}});
         }
     }
-    this->showTabsAction_->setShortcut(toggleSeq);
+    this->toggleTabsAction_->setShortcut(toggleSeq);
 }
 
 bool Notebook::getShowAddButton() const
@@ -1274,7 +1279,7 @@ void Notebook::setLockNotebookLayout(bool value)
 
 void Notebook::addNotebookActionsToMenu(QMenu *menu)
 {
-    menu->addAction(this->showTabsAction_);
+    menu->addAction(this->toggleTabsAction_);
 
     menu->addAction(this->lockNotebookLayoutAction_);
 
@@ -1465,6 +1470,11 @@ void SplitNotebook::toggleOfflineTabs()
                 ? NotebookTabVisibility::AllTabs
                 : NotebookTabVisibility::LiveOnly);
     }
+}
+
+QAction *SplitNotebook::toggleOfflineTabsAction()
+{
+    return this->toggleOfflineTabsAction_;
 }
 
 void SplitNotebook::addNotebookActionsToMenu(QMenu *menu)