+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 ## 2.5.1
 
 - Bugfix: Fixed links without a protocol not being clickable. (#5345)
+- Bugfix: Fixed <kbd>Ctrl</kbd>+<kbd>U</kbd> and "Toggle visibility of tabs" not having the same effect. (#5353)
 
 ## 2.5.0